requirements.txt
pandas>=1.3.0
numpy>=1.21.0
folium>=0.12.0
newsapi-python>=0.2.6
scikit-learn>=0.24.2
spacy>=3.1.0
geopy>=2.2.0
streamlit>=1.10.0
streamlit-folium>=0.7.0
plotly>=5.3.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.1.0/en_core_web_sm-3.1.0.tar.gz
tensorflow>=2.8.0
geopandas>=0.10.0
statsmodels>=0.13.0
xgboost>=1.5.0
prophet>=1.0.0
sklearn-crfsuite>=0.3.6
matplotlib>=3.5.0
seaborn>=0.11.0 